refactor: Using mybatis-plus

+++ server/java/carproject-server/src/main/java/cn/xuanzhi33/carprojectserver/pojo/PagedSensorDataVO.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 @Data
 public class PagedSensorDataVO {
-    private int page;
-    private int pageSize;
-    private int total;
+    private long page;
+    private long pageSize;
+    private long total;
     private List<SensorData> data;
 }
